+++ i18n.js
@@ -443,7 +443,7 @@
             notify_start:                    'Event Start',
             notify_creation:                 'Creation & Updates',
             notify_calamity_reminder:        '10 min reminder (Rounds)',
-            notify_gvg_daily_tasks_desc:     'Daily Tasks breakdown (10:30 UTC)',
+            notify_gvg_daily_tasks_desc:     'Daily Tasks breakdown (11:00 UTC)',
             notify_gvg_pvp_desc:             'Enable Day 6',
             notify_svs_pvp_desc:             'Enable Day 6',
             notify_svs_garrison_desc:        'Garrison reminder',